Fixed #68
ショートカットキー実行時の処理を修正
